Move SDK storybook themes into embedding-sdk-shared
