+++ eslint.config.mjs
@@ -807,7 +807,10 @@ const configs = [
     },
   },
   {
-    files: ["frontend/src/embedding-sdk-bundle/test/**/*"],
+    files: [
+      "frontend/src/embedding-sdk-bundle/test/**/*",
+      "frontend/src/embedding-sdk-shared/test/storybook-themes.ts",
+    ],
     rules: {
       "metabase/no-color-literals": "off",
     },
